move installation back to ./cmd/installer
