+++ cmd/state-installer/internal/installer/installation.go
@@ -1,4 +1,4 @@
-package installation
+package installer
 
 import (
 	"fmt"
