+++ cmd/state-installer/internal/installer/installation_darwin.go
@@ -1,6 +1,6 @@
 // +build darwin
 
-package installation
+package installer
 
 // InstallSystemFiles installs files in the /Application directory
 func InstallSystemFiles(installDir string) error {
